Moved the log for the scoring updates functionality into a `Journal` class to keep this logic separate from the `ScoringSet`.
