+++ js/adapt-contrib-scoring.js
@@ -23,6 +23,7 @@ import IntersectionSet from './IntersectionSet';
 import LifecycleSet from './LifecycleSet';
 import ScoringSet from './ScoringSet';
 import Objective from './Objective';
+import Journal from './Journal';
 import State from './State';
 import StateModels from './StateModels';
 import StateSetModelChildren from './StateSetModelChildren';
@@ -41,6 +42,7 @@ export {
   LifecycleSet,
   ScoringSet,
   Objective,
+  Journal,
   State,
   StateSetModelChildren,
   StateModels
